Increase penalty to ensure constraints are satisfied
Previously, the ground state did not satisfy the constraints for 5 out
of 8 of the feature selection sizes.  This is resolved by increasing
the penalty in the k-out-of-n component of the energy function.

In order to make the specification of the penalty value more general,
a function has been added to compute the maximum change in energy when
flipping a single variable.