Correction to interaction coefficients
Following the original paper from Nguyen et al. (2014), for a
symmetric representation of the Q matrix, the interaction terms should
be given by 1/2 (CMI_01 + CMI_10), whereas previously the code was
using the sum CMI_01 + CMI_10.

Verified that this change does not affect the ground state for the
titanic data set.